+++ xops-fe/src/App.vue
@@ -3,6 +3,7 @@
 
 <template>
 <router-view></router-view>
+
 </template>
 
 <style scoped>