+++ xops-fe/src/views/error-page/404.vue
@@ -1,4 +1,6 @@
 <template>
+
+  
     <div>
       <el-result icon="warning" title="404" sub-title="怎么回事，页面不见啦~">
         <template #extra>