+++ xops-fe/CHANGELOG.md
@@ -22,3 +22,20 @@
 Add .env.development 、.env.production
 
 Add src/types/env.d.ts
+
+# [0.1.0-beta.4] - 2023年10月9日
+
+- change model version
+
+ ```js
+ @types/node              ^20.5.6  →  ^20.8.3
+ @vitejs/plugin-vue        ^4.3.3  →   ^4.4.0
+ element-plus             ^2.3.10  →  ^2.3.14
+ unocss                   ^0.55.2  →  ^0.56.5
+ unplugin-icons           ^0.16.5  →  ^0.17.0
+ unplugin-vue-components  ^0.25.1  →  ^0.25.2
+ vite                      ^4.4.9  →  ^4.4.11
+ vue-router                ^4.2.4  →   ^4.2.5
+ vue-tsc                   ^1.8.8  →  ^1.8.18
+ ```
+